% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordevok.R
\name{affinity}
\alias{affinity}
\title{Measuring affinity within two individuals.}
\usage{
affinity(subject1, subject2)
}
\arguments{
\item{subject1}{Evoked vector of the individual 1.}

\item{subject2}{Evoked vector of the individual 2.}
}
\value{
\code{affinity} returns a numeric measure of affinity between the vectors
inputed.
}
\description{
Under Free Word Association Technique, this function measures the affinity
between the evoked vectors of two individuals.
}
\details{
Under the criteria of the Free Word Association Technique, we propose a
coefficient to calculate the affinity (similiraty) between two vectors of evocations.
Details on the affinity coefficient can be found in
\url{http://wesleyhenriquesp.wixsite.com/rwordevok/affinity}.

The maximum number of evocations allowed to a vector is controlled by the length
of the vectors. Therefore, the vectors \code{subject1} and \code{subject2}
must have equal size. If an individual pronounces less than the maximum \code{N},
the vector must be structured so that n evocations pronounced are in the first \code{n}
positions of the vector according to their importance to the individual and the
remaining \code{N - n} positions must be filled with NA observations.
}
\examples{
#Creating evocation's vectors:
Murillo = c("Regression Analysis","Multivariate Statistics",
"General Statistiscs", "Experiment Planning", "Sampling")
Ingrid = c("Regression Analysis","Multivariate Statistics",
"Temporal Series","Statistical Quality Control", NA)
affinity(Murillo,Ingrid)
}
\references{
Pereira, W. H. S. (2017).
}
\author{
Wesley H. S. Pereira (ICEx/UFMG), Denise Duarte (ICEx/UFMG), Rodrigo B. Ribeiro (IMPA), Gilvan Guedes (CEDEPLAR/UFMG).
}
