% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordevok.R
\name{split_list_wordevok}
\alias{split_list_wordevok}
\title{Cria subgrupos de uma lista de acordo com os grupos informados.}
\usage{
split_list_wordevok(list, groups)
}
\arguments{
\item{list}{Uma lista de arestas.}

\item{groups}{Uma lista contendo os membros da respectiva comunidade.
....}
}
\value{
\code{split_list_wordevok} retorna uma lista contendo em cada dimensão
as ligações por comunidade. Assim, a função ignora as ligações entre comunidades, deixando apenas as
ligações entre membros de uma mesma comunidade.
...
}
\description{
Esta função cria sublistas de adjacência mediante à informação das
comunidades.
}
\examples{
data("preferences")
mtx = wordevok_meaning_list(noloop.pref)
g = graph_from_data_frame(mtx[,c(2,3)], directed = FALSE)
c = cluster_louvain(g)
lst = split_list_wordevok(mtx,c)
}
\author{
Wesley Henrique Silva Pereira
}
