% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordevok.R
\name{telp_wordevok_quad}
\alias{telp_wordevok_quad}
\title{Divide as evocações por seus respectivos quadrantes de Vergès segundo o método selecionado.}
\usage{
telp_wordevok_quad(mtwfile, method = "mean")
}
\arguments{
\item{mtwfile}{Um objeto da classe \code{Multiple-TELP-wordevok}, gerado pela função \code{telp_wordevok}.}

\item{method}{Criterio de discriminação dos quadrantes.
....}
}
\value{
\code{telp_wordevok_quad} retorna uma lista contendo os quadrantes de Vergès
da representação social. Cada dimensão do objeto resultante representa uma comunidade, e cada subdimensão
representa um quadrante daquela comunidade.
...
}
\description{
Divide as evocações por seus respectivos quadrantes de Vergès segundo o método selecionado.
}
\examples{
data("preferences")
mtx = wordevok_affinity_adjacency(noloop.pref)
g = graph_from_adjacency_matrix(mtx,mode = "undirected", weighted = TRUE,
add.rownames = "names")
c = cluster_louvain(g, weights = E(g)$weight)
t = telp_wordevok(noloop.pref,c)
twq = telp_wordevok_quad(t,"mean")
}
\author{
Wesley Henrique Silva Pereira
}
