% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordevok.R
\name{wordevok_adjacency_to_list}
\alias{wordevok_adjacency_to_list}
\title{Transforma uma matriz de adjacências em uma lista de arestas.}
\usage{
wordevok_adjacency_to_list(adjacency)
}
\arguments{
\item{adjacency}{Uma matriz de adjacencias.
...}
}
\value{
\code{wordevok_adjacency_to_list} retorna uma lista de adjacências
conforme fornecida pela matriz de adjacências original. É importante ressaltar
que essa transformação ignora as limitações de se representar as conexões do
grafo através de uma lista de adjacência (como por exemplo a incapacidade de
represnetar vértices isolados). As modificações necessárias ficam
a cargo do usuário.
...
}
\description{
Esta função transforma a matriz de adjacências em uma lista de arestas.
}
\examples{
data("preferences")
mtx = wordevok_affinity_adjacency(noloop.pref)
lst = wordevok_adjacency_to_list(mtx)
}
\author{
Wesley Henrique Silva Pereira
}
