% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordevok.R
\name{wordevok_affinity_adjacency}
\alias{wordevok_affinity_adjacency}
\title{Construction of the adjacency matrix by affinity coefficient.}
\usage{
wordevok_affinity_adjacency(wordevok)
}
\arguments{
\item{wordevok}{A object of wordevok class.}
}
\value{
\code{wordevok_affinity_adjacency} returns the adjacency matrix in which each
line and each column are one indiviual
and each cell is the affinity coefficient between the line's individual and
column's individual.
}
\description{
This function constructs the adjacency matrix by affinity coefficient.
}
\examples{
data("preferences")
mtx = wordevok_affinity_adjacency(noloop.pref)
mtx
}
\author{
Wesley Henrique Silva Pereira
}
