% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordevok.R
\name{wordevok_comm_submatrix}
\alias{wordevok_comm_submatrix}
\title{Extract the submatrix of adjacency matrix under founded communities.}
\usage{
wordevok_comm_submatrix(adjacency, groups, laplacian = FALSE)
}
\arguments{
\item{adjacency}{A adjacency matrix.}

\item{groups}{Uma lista contendo os membros da respectiva comunidade.}

\item{laplacian}{...
....}
}
\value{
\code{wordevok_comm_submatrix} ...
...
}
\description{
....
}
\examples{
data("preferences")
mtx = wordevok_affinity_adjacency(noloop.pref)
g = graph_from_adjacency_matrix(mtx,mode = "undirected", weighted = TRUE,
add.rownames = "names")
c = cluster_louvain(g,weights = E(g)$weight)
sub = wordevok_comm_submatrix(mtx,c,laplacian = FALSE)
}
\author{
Wesley Henrique Silva Pereira
}
