% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordevok.R
\name{wordevok_comm_subsets}
\alias{wordevok_comm_subsets}
\title{Extrai uma lista contendo as evocações das comunidades do objeto wordevok.}
\usage{
wordevok_comm_subsets(wordevok, groups)
}
\arguments{
\item{wordevok}{Um objeto da classe wordevok.}

\item{groups}{Uma lista contendo os membros da respectiva comunidade.
....}
}
\value{
\code{wordevok_comm_subsets} Retorna uma lista onde em cada uma das
dimensões são alocados os subconjuntos do objeto \code{wordevok} orginal
de acordo com as comunidades informadas.
...
}
\description{
Essa função extrai, mediante a informação dos membros de cada comunidade
encontrada por um método de agrupamento em redes, uma lista contendo os
subconjuntos de evocações de cada uma das comunidades informadas.
}
\examples{
data("preferences")
mtx = wordevok_affinity_adjacency(noloop.pref)
g = graph_from_adjacency_matrix(mtx,mode = "undirected", weighted = TRUE,
add.rownames = "names")
c = cluster_louvain(g,weights = E(g)$weight)
sub = wordevok_comm_subsets(noloop.pref,c)
}
\author{
Wesley Henrique Silva Pereira
}
