% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordevok.R
\name{wordevok_intersection_adjacency}
\alias{wordevok_intersection_adjacency}
\title{Construction of the adjacency matrix by intersection.}
\usage{
wordevok_intersection_adjacency(wordevok)
}
\arguments{
\item{wordevok}{A wordevok class object.}
}
\value{
\code{wordevok_intersection_adjacency} returns the adjacency matrix in which
each line and each column are one
indiviual and each cell is the common evocation's number between the line's
individual and column's individual.
}
\description{
This function constructs the adjacency matrix by the number of
evocations in common.
}
\examples{
data("preferences")
mtx = wordevok_intersection_adjacency(noloop.pref)
mtx
}
\author{
Wesley Henrique Silva Pereira
}
