% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordevok.R
\name{wordevok_laplacian_adjacency}
\alias{wordevok_laplacian_adjacency}
\title{Construction of the adjacency matrix by laplacian method.}
\usage{
wordevok_laplacian_adjacency(wordevok, normalized = FALSE)
}
\arguments{
\item{wordevok}{A wordevok class object.}

\item{normalized}{Logical argument. If \code{normalized = TRUE}, the output will be
normalized.}
}
\value{
\code{wordevok_laplacian_adjacency} returns the adjacency matrix in which each
line and each column are an
evocation and each cell takes on -1 if the line's individual and column's
individual have common evocations. The diagonal of the matrix equals the number of
edges incident on the vertex.
}
\description{
This function constructs the adjacency matrix by laplacian method.
}
\examples{
data("preferences")
mtx = wordevok_laplacian_adjacency(noloop.pref)
mtx
}
\author{
Wesley Henrique Silva Pereira
}
