% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordevok.R
\name{wordevok_meaning_adjacency}
\alias{wordevok_meaning_adjacency}
\title{Construction of the adjacency matrix by meaning weighted method.}
\usage{
wordevok_meaning_adjacency(wordevok)
}
\arguments{
\item{wordevok}{A wordevok class object.}
}
\value{
\code{wordevok_meaning_adjacency} returns the adjacency matrix in which each
line and each column are an evocation
and each cell takes on 1 if the line's individual and column's individual have
common evocations. Otherwise, the cells takes on 0.
}
\description{
This function constructs the adjacency matrix by meaning weighted method.
}
\examples{
data("preferences")
mtx = wordevok_meaning_adjacency(noloop.pref)
mtx
}
\author{
Wesley Henrique Silva Pereira
}
