% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordevok.R
\name{wordevok_meaning_list}
\alias{wordevok_meaning_list}
\title{Construction of the list of edges by meaning weighted method.}
\usage{
wordevok_meaning_list(wordevok, output = "Complete")
}
\arguments{
\item{wordevok}{A wordevok class object.}

\item{output}{Type of output to be generated. See details to more information.}
}
\value{
\code{wordevok_meaning_list} returns a list containing the edges and their
respective weights.
}
\description{
This function constructs the list of edges by meaning weighted method.
}
\details{
The output argument of the function x must be:

- "Complete": the object contains the identifier of the individual
generating the edge, the vertices bordering the edge and the weights
assigned to the edges.

- "Edgelist": the same elements of the "Complete" output, except the
identifier of the individual generator.

- "Simplify": this output contains only the unique edges and their weights.
The weights associated with these represent the sum of the
weights of all your duplicates.
}
\examples{
data("preferences")
lst = wordevok_meaning_list(noloop.pref)
lst
}
\author{
Wesley Henrique Silva Pereira
}
