% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordevok.R
\name{wordevok_quad_class}
\alias{wordevok_quad_class}
\title{Calcula as coordenadas da relevância para os signficados dos quadrantes de
uma determinada comunidade.}
\usage{
wordevok_quad_class(quad, classe)
}
\arguments{
\item{quad}{Uma subdimensão de um objeto da classe \code{TELP-wordevok-quad} gerado pela função
telp-wordevok-quad.}

\item{classe}{Um documento listando as classes às quais pertencem as evocações
únicas do wordevok. Veja os detalhes para mais informações.
....}
}
\value{
\code{wordevok_quad_class} Retorna as coordenadas de relevância relativa
de um determinado significado para um quadrante de uma comunidade.
...
}
\description{
Calcula as coordenadas da relevância para os signficados dos quadrantes de
uma determinada comunidade.
}
\examples{
data("preferences")
mtx = wordevok_affinity_adjacency(noloop.pref)
g = graph_from_adjacency_matrix(mtx,mode = "undirected", weighted = TRUE,
add.rownames = "names")
c = cluster_louvain(g, weights = E(g)$weight)
t = telp_wordevok(noloop.pref,c)
twq = telp_wordevok_quad(t,"mean")

# Macro-groups

u = noloop.pref$Evocations
mg = c("Optional","Optional","Fifth","Fifth","Second","First","Optional",
      "Eigth","Fifth","Third","Second","Forth","Optional","Forth","Sixth",
      "Third","Optional","Forth","Optional","Sixth")
u = data.frame(cbind(u,mg))
wqc= wordevok_quad_class(twq[[1]][[1]],u)
}
\author{
Wesley Henrique Silva Pereira
}
