% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordevok.R
\name{wordevok_radar_gg}
\alias{wordevok_radar_gg}
\title{Adequa as coordenadas para a utilização dos radares criados usando o \code{ggplot2}.}
\usage{
wordevok_radar_gg(radar_attr)
}
\arguments{
\item{radar_attr}{Um objeto da classe \code{TELP-wordevok-radar} criado pela
função \core{wordevok_radar_attr}.
....}
}
\value{
\code{wordevok_radar_gg} retorna uma lista onde cada dimensão representa
o núcleo central da comunidade em questão bem como as relevâncias relativas de
cada conceito na mesma.
...
}
\description{
Adequa as coordenadas para a utilização dos radares criados usando o \code{ggplot2}.
}
\examples{
data("preferences")
mtx = wordevok_affinity_adjacency(noloop.pref)
g = graph_from_adjacency_matrix(mtx,mode = "undirected", weighted = TRUE,
add.rownames = "names")
c = cluster_louvain(g, weights = E(g)$weight)
t = telp_wordevok(noloop.pref,c)
twq = telp_wordevok_quad(t,"mean")

# Macro-groups

u = noloop.pref$Evocations
mg = c("Optional","Optional","Fifth","Fifth","Second","First","Optional",
      "Eigth","Fifth","Third","Second","Forth","Optional","Forth","Sixth",
      "Third","Optional","Forth","Optional","Sixth")
u = data.frame(cbind(u,mg))
wra= wordevok_radar_attr(twq,u)
wrg= wordevok_radar_gg(wra)
}
\author{
Wesley Henrique Silva Pereira
}
