% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordevok.R
\name{wordevok_radar_plot}
\alias{wordevok_radar_plot}
\title{Gera os gráficos de radar para o pensamento coletivo das comunidades usando \code{ggplot2}.}
\usage{
wordevok_radar_plot(radar_gg)
}
\arguments{
\item{radar_gg}{Um objeto da classe \code{wordovok-radar-gg} criado pela
função \core{wordevok_radar_gg}.
....}
}
\value{
\code{wordevok_radar_plot} retorna uma lista onde cada dimensão representa
um gráfico de radar do núcleo central da comunidade em questão.
...
}
\description{
....
}
\examples{
data("preferences")
mtx = wordevok_affinity_adjacency(noloop.pref)
g = graph_from_adjacency_matrix(mtx,mode = "undirected", weighted = TRUE,
add.rownames = "names")
c = cluster_louvain(g, weights = E(g)$weight)
t = telp_wordevok(noloop.pref,c)
twq = telp_wordevok_quad(t,"mean")

# Macro-groups

u = noloop.pref$Evocations
mg = c("Optional","Optional","Fifth","Fifth","Second","First","Optional",
      "Eigth","Fifth","Third","Second","Forth","Optional","Forth","Sixth",
      "Third","Optional","Forth","Optional","Sixth")
u = data.frame(cbind(u,mg))
wra = wordevok_radar_attr(twq,u)
wrg = wordevok_radar_gg(wra)
wrp = wordevok_radar_plot(wrg)
wrp[[1]]
}
\author{
Wesley Henrique Silva Pereira
}
